#1958 – moved salts and solvents logic to hook
